src/modules/text-editor/src/services/transforms/italic.transform.ts
import { Injectable } from '@angular/core';

import { ExecCommand } from '../../text-editor.interface';
import { Transform } from '../../types/transform';
import { BaseTransform } from './base.transform';

@Injectable()
export class ItalicTransform extends BaseTransform implements Transform {

  toggle(): void {
    this.execCommand(ExecCommand.Italic);
  }

  get value(): boolean {
    if (!this.isSelectionExist) {
      return false;
    }

    return document.queryCommandState(ExecCommand.Italic);
  }

}
